source files related to dynamic obstacle detection moved into a separat folder
