+++ CMakeLists.txt
@@ -133,12 +133,12 @@ add_library(costmap_converter 	src/costmap_to_polygons.cpp
 				src/costmap_to_polygons_concave.cpp
 				src/costmap_to_lines_convex_hull.cpp
 				src/costmap_to_lines_ransac.cpp
-				src/costmap_to_dynamic_obstacles.cpp
-				src/background_subtractor.cpp
-				src/blob_detector.cpp
-				src/Ctracker.cpp
-				src/Kalman.cpp
-				src/HungarianAlg.cpp)
+				src/costmap_to_dynamic_obstacles/costmap_to_dynamic_obstacles.cpp
+				src/costmap_to_dynamic_obstacles/background_subtractor.cpp
+				src/costmap_to_dynamic_obstacles/blob_detector.cpp
+				src/costmap_to_dynamic_obstacles/Ctracker.cpp
+				src/costmap_to_dynamic_obstacles/Kalman.cpp
+				src/costmap_to_dynamic_obstacles/HungarianAlg.cpp)
 target_link_libraries(costmap_converter ${catkin_LIBRARIES} )  
 
 ## Declare a cpp executable
