+++ package.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0"?>
-<package>
+<package format="2">
   <name>costmap_converter</name>
   <version>0.0.5</version>
   <description>
@@ -9,27 +9,27 @@
   <maintainer email="christoph.roesmann@tu-dortmund.de">Christoph Rösmann</maintainer>
 
   <author email="christoph.roesmann@tu-dortmund.de">Christoph Rösmann</author>
-  <author email="otniel.rinaldo@tu-dortmund.de">Otniel Rinaldo</author>
+  <author email="franz.albers@tu-dortmund.de">Franz Albers</author>
+  <author email="otniel.rinaldo@tu-dortmund.de">Otniel Rinaldo</author> 
 
   <license>BSD</license>
 
   <url type="website">http://wiki.ros.org/costmap_converter</url>
   
   <buildtool_depend>catkin</buildtool_depend>
-  <build_depend>geometry_msgs</build_depend>
-  <build_depend>roscpp</build_depend>
-  <build_depend>std_msgs</build_depend>
-  <build_depend>costmap_2d</build_depend>
-  <build_depend>dynamic_reconfigure</build_depend>
-  <build_depend>pluginlib</build_depend>
-  <build_depend>cv_bridge</build_depend>
-  <run_depend>geometry_msgs</run_depend>
-  <run_depend>roscpp</run_depend>
-  <run_depend>std_msgs</run_depend>
-  <run_depend>costmap_2d</run_depend>
-  <run_depend>dynamic_reconfigure</run_depend>
-  <run_depend>pluginlib</run_depend>
-  <run_depend>cv_bridge</run_depend>
+
+  <build_depend>message_generation</build_depend>
+
+  <build_export_depend>message_runtime</build_export_depend>
+  <exec_depend>message_runtime</exec_depend>
+
+  <depend>geometry_msgs</depend>
+  <depend>roscpp</depend>
+  <depend>std_msgs</depend>
+  <depend>costmap_2d</depend>
+  <depend>dynamic_reconfigure</depend>
+  <depend>pluginlib</depend>
+  <depend>cv_bridge</depend>
 
   <!-- The export tag contains other, unspecified, tags -->
   <export>
