+++ plugins.xml
@@ -31,7 +31,7 @@
   
   <class type="costmap_converter::CostmapToDynamicObstacles" base_class_type="costmap_converter::BaseCostmapToPolygons">
   <description>
-      This class converts costmap2d obstacles into a vector dynamic obstacles. 
+      This class detects and tracks obstacles from a sequence of costmaps. 
   </description>
   </class>
 </library>
